Fixes #3077 Correctly rollback multiple versions consecutively (#5242)
