+++ inc/Engine/Plugin/UpdaterSubscriber.php
@@ -1,12 +1,15 @@
 <?php
 namespace WP_Rocket\Engine\Plugin;
 
-use WP_Rocket\Event_Management\Subscriber_Interface;
+use Plugin_Upgrader;
+use Plugin_Upgrader_Skin;
+use WP_Rocket\Event_Management\Event_Manager;
+use WP_Rocket\Event_Management\Event_Manager_Aware_Subscriber_Interface;
 
 /**
  * Manages the plugin updates.
  */
-class UpdaterSubscriber implements Subscriber_Interface {
+class UpdaterSubscriber implements Event_Manager_Aware_Subscriber_Interface {
 	use UpdaterApiTools;
 
 	/**
@@ -64,6 +67,13 @@ class UpdaterSubscriber implements Subscriber_Interface {
 	protected $cache_transient_name = 'wp_rocket_update_data';
 
 	/**
+	 * The WordPress Event Manager
+	 *
+	 * @var Event_Manager;
+	 */
+	protected $event_manager;
+
+	/**
 	 * Constructor
 	 *
 	 * @param array $args {
@@ -85,6 +95,15 @@ class UpdaterSubscriber implements Subscriber_Interface {
 
 	/**
 	 * {@inheritdoc}
+	 *
+	 * @param Event_Manager $event_manager The WordPress Event Manager.
+	 */
+	public function set_event_manager( Event_Manager $event_manager ) {
+		$this->event_manager = $event_manager;
+	}
+
+	/**
+	 * {@inheritdoc}
 	 */
 	public static function get_subscribed_events() {
 		return [
@@ -93,6 +112,7 @@ class UpdaterSubscriber implements Subscriber_Interface {
 			'deleted_site_transient'                => 'maybe_delete_rocket_update_data_cache',
 			'wp_rocket_loaded'                      => 'maybe_force_check',
 			'auto_update_plugin'                    => [ 'disable_auto_updates', 10, 2 ],
+			'admin_post_rocket_rollback'            => 'rollback',
 		];
 	}
 
@@ -404,5 +424,84 @@ class UpdaterSubscriber implements Subscriber_Interface {
 	 */
 	public function delete_rocket_update_data_cache() {
 		delete_site_transient( $this->cache_transient_name );
+	}
+
+	/**
+	 * Do the rollback
+	 *
+	 * @since 2.4
+	 */
+	public function rollback() {
+		if ( ! isset( $_GET['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_GET['_wpnonce'] ), 'rocket_rollback' ) ) {
+			wp_nonce_ays( '' );
+		}
+
+		/**
+		 * Fires before doing the rollback
+		 */
+		do_action( 'rocket_before_rollback' );
+
+		$plugin_transient = get_site_transient( 'update_plugins' );
+		$plugin_folder    = plugin_basename( dirname( WP_ROCKET_FILE ) );
+		$plugin           = $plugin_folder . '/' . basename( WP_ROCKET_FILE );
+
+		$plugin_transient->response[ $plugin ] = (object) [
+			'slug'        => $plugin_folder,
+			'new_version' => WP_ROCKET_LASTVERSION,
+			'url'         => 'https://wp-rocket.me',
+			'package'     => sprintf( 'https://wp-rocket.me/%s/wp-rocket_%s.zip', get_rocket_option( 'consumer_key' ), WP_ROCKET_LASTVERSION ),
+		];
+
+		$this->event_manager->remove_callback( 'pre_set_site_transient_update_plugins', [ $this, 'maybe_add_rocket_update_data' ] );
+
+		set_site_transient( 'update_plugins', $plugin_transient );
+
+		require_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
+
+		// translators: %s is the plugin name.
+		$title         = sprintf( __( '%s Update Rollback', 'rocket' ), WP_ROCKET_PLUGIN_NAME );
+		$nonce         = 'upgrade-plugin_' . $plugin;
+		$url           = 'update.php?action=upgrade-plugin&plugin=' . rawurlencode( $plugin );
+		$upgrader_skin = new Plugin_Upgrader_Skin( compact( 'title', 'nonce', 'url', 'plugin' ) );
+		$upgrader      = new Plugin_Upgrader( $upgrader_skin );
+
+		add_filter( 'update_plugin_complete_actions', [ $this, 'rollback_add_return_link' ] );
+		rocket_put_content( WP_CONTENT_DIR . '/advanced-cache.php', '' );
+
+		$upgrader->upgrade( $plugin );
+
+		wp_die(
+			'',
+			// translators: %s is the plugin name.
+			esc_html( sprintf( __( '%s Update Rollback', 'rocket' ), WP_ROCKET_PLUGIN_NAME ) ),
+			[
+				'response' => 200,
+			]
+		);
+	}
+
+	/**
+	 * After a rollback has been done, replace the "return to" link by a link pointing to WP Rocket's tools page.
+	 * A link to the plugins page is kept in case the plugin is not reactivated correctly.
+	 *
+	 * @since  3.2.4
+	 *
+	 * @param  array $update_actions Array of plugin action links.
+	 * @return array                 The array of links where the "return to" link has been replaced.
+	 */
+	public function rollback_add_return_link( $update_actions ) {
+		if ( ! isset( $update_actions['plugins_page'] ) ) {
+			return $update_actions;
+		}
+
+		$update_actions['plugins_page'] = sprintf(
+			/* translators: 1 and 3 are link openings, 2 is a link closing. */
+			__( '%1$sReturn to WP Rocket%2$s or %3$sgo to Plugins page%2$s', 'rocket' ),
+			'<a href="' . esc_url( admin_url( 'options-general.php?page=' . WP_ROCKET_PLUGIN_SLUG ) . '#tools' ) . '" target="_parent">',
+			'</a>',
+			'<a href="' . esc_url( admin_url( 'plugins.php' ) ) . '" target="_parent">'
+		);
+
+		return $update_actions;
 	}
 }
