+++ inc/deprecated/3.11.php
@@ -199,3 +199,88 @@ function rocket_remove_partial_purge_hooks() {
 
 	remove_filter( 'rocket_clean_files', 'rocket_clean_files_users' );
 }
+
+/**
+ * Do the rollback
+ *
+ * @since 3.11.5 deprecated
+ * @since 2.4
+ */
+function rocket_rollback() {
+	_deprecated_function( __FUNCTION__ . '()', '3.11.5' );
+	if ( ! isset( $_GET['_wpnonce'] ) || ! wp_verify_nonce( sanitize_key( $_GET['_wpnonce'] ), 'rocket_rollback' ) ) {
+		wp_nonce_ays( '' );
+	}
+
+	/**
+	 * Fires before doing the rollback
+	 */
+	do_action( 'rocket_before_rollback' );
+
+	$plugin_transient = get_site_transient( 'update_plugins' );
+	$plugin_folder    = plugin_basename( dirname( WP_ROCKET_FILE ) );
+	$plugin           = $plugin_folder . '/' . basename( WP_ROCKET_FILE );
+
+	$plugin_transient->response[ $plugin ] = (object) [
+		'slug'        => $plugin_folder,
+		'new_version' => WP_ROCKET_LASTVERSION,
+		'url'         => 'https://wp-rocket.me',
+		'package'     => sprintf( 'https://wp-rocket.me/%s/wp-rocket_%s.zip', get_rocket_option( 'consumer_key' ), WP_ROCKET_LASTVERSION ),
+	];
+
+	set_site_transient( 'update_plugins', $plugin_transient );
+
+	require_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
+
+	// translators: %s is the plugin name.
+	$title         = sprintf( __( '%s Update Rollback', 'rocket' ), WP_ROCKET_PLUGIN_NAME );
+	$nonce         = 'upgrade-plugin_' . $plugin;
+	$url           = 'update.php?action=upgrade-plugin&plugin=' . rawurlencode( $plugin );
+	$upgrader_skin = new Plugin_Upgrader_Skin( compact( 'title', 'nonce', 'url', 'plugin' ) );
+	$upgrader      = new Plugin_Upgrader( $upgrader_skin );
+
+	remove_filter( 'site_transient_update_plugins', 'rocket_check_update', 1 );
+	add_filter( 'update_plugin_complete_actions', 'rocket_rollback_add_return_link' );
+	rocket_put_content( WP_CONTENT_DIR . '/advanced-cache.php', '' );
+
+	$upgrader->upgrade( $plugin );
+
+	wp_die(
+		'',
+		// translators: %s is the plugin name.
+		esc_html( sprintf( __( '%s Update Rollback', 'rocket' ), WP_ROCKET_PLUGIN_NAME ) ),
+		[
+			'response' => 200,
+		]
+	);
+}
+
+/**
+ * After a rollback has been done, replace the "return to" link by a link pointing to WP Rocket's tools page.
+ * A link to the plugins page is kept in case the plugin is not reactivated correctly.
+ *
+ * @since 3.11.5 deprecated
+ * @since  3.2.4
+ * @author Grégory Viguier
+ * @author Arun Basil Lal
+ *
+ * @param  array $update_actions Array of plugin action links.
+ * @return array                 The array of links where the "return to" link has been replaced.
+ */
+function rocket_rollback_add_return_link( $update_actions ) {
+	_deprecated_function( __FUNCTION__ . '()', '3.11.5' );
+
+	if ( ! isset( $update_actions['plugins_page'] ) ) {
+		return $update_actions;
+	}
+
+	$update_actions['plugins_page'] = sprintf(
+		/* translators: 1 and 3 are link openings, 2 is a link closing. */
+		__( '%1$sReturn to WP Rocket%2$s or %3$sgo to Plugins page%2$s', 'rocket' ),
+		'<a href="' . esc_url( admin_url( 'options-general.php?page=' . WP_ROCKET_PLUGIN_SLUG ) . '#tools' ) . '" target="_parent">',
+		'</a>',
+		'<a href="' . esc_url( admin_url( 'plugins.php' ) ) . '" target="_parent">'
+	);
+
+	return $update_actions;
+}
